Keep Service Worker (Registration) Objects reference counted.
Keep the Service Worker Registration Object reference counted in the
Service Worker Registration wrappable object, so that it can be
referenced, even when it has been terminated.  This avoids crashing
when trying to unregister a worker that already terminated.

Keep the Service Worker Object reference counted in the Service
Worker wrappable object, so that we can test if the global scope of
the service worker still exists before posting a message.  This
avoids crashing when attempting to post a message to a terminated service worker.

b/228905127
b/228909717

Change-Id: Id2af67cbe32a097e5c3c6b098797c97d3ccc6f6f
NOKEYCHECK=True
GitOrigin-RevId: 159fd650f8c90e1e8753c858c4abc2633707c58b

Former-commit-id: a05a7f3e1f6c46705e000705478964fa7147c867