Fix for creation URL and API base URL of environment settings.
This fixes the creation URL and API base URL of environment settings to
follow the spec more accurately and reflect URL redirection.

This logic was broken since go/cobalt-cl/220581.

The dynamic resolution of base_url conflicted with the setup of the
websocket unit tests, so this CL also includes a refactor of that to
the new TestWithJavaScript helper, and the relevant changes to
StubWindow to allow the relevant test initialization.

b/241471984

Change-Id: I28711002252948cde0364df6b4f172f267ab78d1
NOKEYCHECK=True
GitOrigin-RevId: dca6ff4a36901a893ad5048fa54cf7b8b3bb8fa0

Former-commit-id: 0a7a6b8d0dc593f7459ab8fd40f2bd5e6c3d6965