Fix behavior with syntax error in synchronous JS in service worker.
When the worker script execution fails, this now dispatches the error
event to the client from the correct thread, fixing a run-time assert.

This also ensures that the error message is also reported to the error
event handler, with a test added to the web worker black box test.

b/235873258

Change-Id: I283ed8d8ba416d5e740dd3f3d58cb9dfeedb43c3
NOKEYCHECK=True
GitOrigin-RevId: b34d44b98e8427081264739189a4d5484bf91476

Former-commit-id: 68a37b8931ad4ee2ff3b831febc2fe3b61e99f59