Fix value for 'origin' in the environment settings object.
The value for the 'origin' property of the  environment settings object
was already returning the creation URL. As a result, SOP/CORS checks
against it would yield incorrect results when redirects were
encountered.

This implements the value for 'origin' following the spec, for window
and worker global scopes.

b/241471984

Change-Id: Ia6c8a90047a724024b51109716814c933d43e24a
NOKEYCHECK=True
GitOrigin-RevId: 25da28801d94eb247eae5e4de94b6c6393f65e01

Former-commit-id: 89bf46c097fe283e5fafdcf0fc21cccd04086176