Fix race condition when new WebModule is created.
This fixes a race condition when callbacks are made from a new WebModule
before the BrowerModule updates its WebModule member variable.

When a callback is called from the WebModule thread, the
web_module_ member of BrowserModule can not be relied on, since that is
updated in a different thread and may not yet be assigned at the time
the callback runs.

The BrowserModule::CreateH5vcc callback that creates the window.h5vcc
property, which interfaces to BrowserModule in a few different ways now
receives the WebModule pointer from the WebModule for which it is
being created, using a trampoline for the
injected_global_object_attributes callbacks passed to the web::Agent.

This also applies to the call to SetDeepLinkTimestamp to report to the
Performance API, which now ensures the right thread is used to reference
the WebModule pointer.

b/244482152

Change-Id: Ifdfc5b03006387cef2ffe4ae80d64ab5c8833759
(cherry picked from commit f75bb6a5dfbf8117fad41c3698fab5fa2ea3846b)
NOKEYCHECK=True
GitOrigin-RevId: e764407667c19c66daa4e302d300ce26b36a18d6

Former-commit-id: 86cf1a9e81665a422ab736002d7b172f48af9938