Add more MessageEvent functionality and tests.
This adds origin, lastEventId, source, and ports to MessageEvent, and
adds a set of related unit tests for MessageEvent and
ExtendableMessageEvent.

To support the unit tests, the TestWithJavaScript helper classes have
been changed to use a virtual accessor for the web::Context instead of
the script::GlobalEnvironment, as that still allows access to the latter
and also allows access to the WindowOrWorkerGlobalScope needed for
testing the 'source' property. To still correctly initialize the window_
for the web::testing tests with the Window global scope a call to
InitializeWindow() had to be added.

Also, the EvaluateScript() overload that returns a
script::ValueHandleHolder::Reference was added to the TestWithJavaScript
helpers to allow the unit tests to test with the data property set to a
JS value.

Also renamed GetWorkerTypes() to GetWebTypes() for the web::testing test
case instantiation.

The above changes somehow also triggered some previously unseen compiler
warnings in script/v8c, and some other places that are also fixed in
this CL.

b/236294349

Change-Id: I01462b92348f6ff3a1d209523621ed0759ed383f
(cherry picked from commit 90af64cbba9cc6860d2ac8cd828fa4a6cf9d05f0)
NOKEYCHECK=True
GitOrigin-RevId: da1c2b8b38d124f1880e61a313ffd89fd50170f5

Former-commit-id: 6c49d9d19e70edef1b75127e95a96f19bd8e7eab