[media] Fix DCHECK() in dav1d video decoder ReportError()
The previous implementation of VideoDecoder::ReportError() assumes that it is always called on the |decoder_thread_| and verifies this by a DCHECK().

However, VideoDecoder::ReportError() can be called by VideoDecoder::WriteInputBuffer() which is always called on the player worker thread, which triggers the DCHECK() when error occurs.

Now the DCHECK() in VideoDecoder::ReportError() is removed, and the |error_cb_| is always called on the player worker thread.

b/250549091

Change-Id: Idf44d60cb8d6d56dd4d3ddbe844599bb18e9dcc0
(cherry picked from commit 6dbdcd7c4ac699b3fe43066be3f7141b18306bf0)
NOKEYCHECK=True
GitOrigin-RevId: acda1520daecfa4f60f8ea13a658f9dcffb6e366

Former-commit-id: cfec1bc87c3568766c3afeceaf47b121d6a5143b