[Android] Update MediaSession state when it changes to none
Always update playbackStateBuilder when updateMediaSessionInternal
is called, and set that playback state into mediaSession. Additionally,
have configureMediaFocus also update the playbackStateBuilder and set
that state to handle cases triggered by suspend/resume. This also
handles the case where mediaSession was null (deactive), where
setMediaSession will create a new mediaSession and set whatever state
is present in playbackStateBuilder.

The prior logic for handling suspend/resume had suspend update the
playbackStateBuilder with the values for resume and *not* set that
state in mediaSession before the mediaSession was deactivated. When
resume occurred, the setMediaSession would then use the values within
playbackStateBuilder on the new mediaSession.

b/255774645

Change-Id: I3510033ffb59ab83ec17e0647da204747df4b8f1
NOKEYCHECK=True
GitOrigin-RevId: e20835c98ebb91cacbbc23350435c3f84fc97c87

Former-commit-id: ed7c7ede7d87fe1bb45489d04b3aff6614496f32