Destroy the image cache before the animated image tracker.
The animated image tracker owns the decoding thread for animated images,
but the image cache owns the images that depend on that thread.

That means that there is a dependency of the image cache on the lifetime
of the animated image tracker, and it should be destroyed first.

It may be better for the image cache to own the animated image tracker
to make that dependency more explicit, but this CL is the minimal change
to the code to fix the bug. A warning is added in comments to highlight
the dependency.

A DCHECK() is added in PostBlockingTask when a task that will never run
is posted.

b/255403883

Change-Id: I42fe40974ee848de7b7b49a3b9f9668edad7dec0
(cherry picked from commit 03c992dac20987fd5015a28dfb79c92d07f0dc4b)
NOKEYCHECK=True
GitOrigin-RevId: 8e070f71354c9fa8b9139e960033e6971307f4b2

Former-commit-id: 452c5f30cfb2c40bdff8839e3f25d1a044dd8ac2