Add extension to be notified when the web app is rendered
Cobalt has a custom Web API "renderPostponed" which is used to delay
rendering of the main web module. This is usually used to show just
the splash screen while ***REMOVED*** is loading assets (thumbnails, etc.).
Once this flag is flipped, and the resultant render tree is processed,
call the new extension to notify it that the web app is shown.

b/246566068

Change-Id: Ie4f3916579feb5f199797b11a55e166e4f704909
NOKEYCHECK=True
GitOrigin-RevId: 4b3ac02773cbac31bf90268d133bf88d3036674e

Former-commit-id: db1da6f6f38b64e4aa9a2ac516a085bdcf88aec4