[mediasource] Refine immediate job
In the previous implementation immediate job is only enabled when
asynchronous reduction is enabled.

Now immediate job can be enabled without asynchronous reduction, and the
old logic was kept to ensure that we can continue to experiment on it.

As a result, the max size for immediate job is reduced from 16 KB to 0,
to disable it by default.

b/223283615

Change-Id: Ie7cc48a6b4f2143194b513ef23582ac753b80f86
NOKEYCHECK=True
GitOrigin-RevId: 0785cf7f02464fc150bd0f4ebc1da9536ec5f243

Former-commit-id: 5b192331ba3c3850b817fae2637ee585620dbac4