Resolve heap use after free for promises.
This resolves heap use after free for promises that result from web
contexts being referenced after they are shutdown.

To make this possible, pass around web::Context instead of
web::EnvironmentSettings,

When a web::Context is unregistered, cleanup any references to the
client in any job, and any promises in jobs refering to the client. This
is implemented in PrepareForClientShutdown.

This also includes a cleanup reducing code duplication for resolving or
rejecting equivalent jobs.

b/264920834

Change-Id: I75fa3f34e10b4737acf5773dfa0ab80caf1ed370
NOKEYCHECK=True
GitOrigin-RevId: a15ed16e5b1bbf57a47bade7bb1b0d8541bce6c6

Former-commit-id: 4d106d82ce1cdca212a5aa4c7ec06a188df20cdb