Make test runner fail faster on errors
This does two things:
Change the regex matching to be less specific, so it's more reliable
in picking up the correct log lines for Webmodule load and Webdriver
connection.
It also adds an exit condition for when Cobalt didn't start
correctly, preventing a 2-minute wait for every failed test.

b/135549281

Change-Id: I39f9b74cdaa44d4ddca8dbce595e626ef8fab481
NOKEYCHECK=True
GitOrigin-RevId: 6d05a66e362e398b04c4007135424850e124f93e

Former-commit-id: 87655f31bc3a0be053c7b8b3f53143cbad7e8f73