[media] Fix typo in VideoSampleInfo ctor
It used to use sizeof(SbMediaAudioSampleInfo) to memset to 0.  It worked
as once VideoSampleInfo::codec is set to kSbMediaVideoCodecNone (0) the
other members won't be used.

Correct it by replacing it with sizeof(SbMediaVideoSampleInfo).

Add a test for VideoSampleInfo ctor, and allows for future expansion to
more tests on VideoSampleInfo.

b/241786349

Change-Id: I2fb767c64100597877fa339fd8730617c515c98e
NOKEYCHECK=True
GitOrigin-RevId: aefd8a593afd9d5389c9c9acf7f9c9c34895afac

Former-commit-id: 08521f3311e0d89b9535033bfc4fa62066c07796