Enable a native build of Crashpad handler
That is, a build without use of the Starboard porting layer.

Support is added for the Linux and Raspi reference ports. After running
gn gen for the loader platform, the Crashpad handler binary can be
built, without Starboard, by building the native_target/crashpad_handler target.

The changes are made so that the Starboard-based Crashpad handler binary
is unchanged and can still be built as before, for now. After Buildbot
is migrated to use the native Crashpad handler binary, this will
be cleaned up and it will only be possible to build the Crashpad handler
binary without Starboard.

Support for Evergreen-AOSP will be in a separate, much smaller change.

b/256660693

Change-Id: Id329c84d4900fd2090f1a6433a0be18cfcd76003
NOKEYCHECK=True
GitOrigin-RevId: f36f1fcd3563a965ae3d5a7e2c050f49929cd9d9

Former-commit-id: 7e71752633a8934690718d49d09fa5275c5b0233