[BG] Avoid player creation of audioless content in background mode
As we remove video stream from the content for background mode, when
playing audioless content, starboard would fail to create the player
with no output and hit DCHECKs. As it doesn't make sense to resume the
player with no output, we don't need to resume the player for audioless
content in background mode.

b/270752899
b/272378829
b/272555615

Change-Id: I002f2fe5e13ad47fafd0368e6cbefd81fd3d7ff1
NOKEYCHECK=True
GitOrigin-RevId: 581eaafbbcdfaf5e7b048e2d9031f019f7cba805

Former-commit-id: 1ff6bb2c82c267f94256a90eaaac7f922564acd3