Move error reporting to WindowOrWorkerGlobalScope.
This moves ReportScriptError and its registration with the JavaScript
global environment from WebModule to WindowOrWorkerGlobalScope, so that
the error event is also dispatched in workers when needed. Also adds
corresponding tests.

This also moves the initial EnableEval and ReportEval configuration for
CSP so that it functions correctly in workers.

b/275741116
b/225037465

Change-Id: Ic86caba993197b9a323022363d0c42a8467caf88
NOKEYCHECK=True
GitOrigin-RevId: f3e7d2e2087a5671009dcdd0cf1089d82e612abc

Former-commit-id: 8606ac4b328f30aac7cb15a5eb9833e558b09425