[media] Refactor OutputModeSupported() (#307)
It was declared as a static method in VideoDecoder, and is now moved
into PlayerComponents:Factory.

VideoDecoder::OutputModeSupported() should be a global method to provide
information on whether a particular output mode is supported.
VideoDecoder is a pure virtual class meant to be inherited by individual
platforms, which isn't the best place to contain OutputModeSupported().

For example:
1. A platform may implement two different VideoDecoder subclasses to
   support h264 and vp9, and VideoDecoder::OutputModeSupported() doesn't
   belong to either of them.
2. A VideoDecoder subclass may be shared by multiple platforms, and we
   cannot put VideoDecoder::OutputModeSupported() into its source file
   as OutputModeSupported() might not be shared by these platforms.

As a result, VideoDecoder::OutputModeSupported() is often placed inside
the implementation file of PlayerComponents::Factory.  Each platform has
its own PlayerComponents::Factory implementation, and is the best place
to provide platform specific information for all media filters.  Thus
OutputModeSupported() is moved from VideoDecoder to
PlayerComponents::Factory.

b/175072281
Former-commit-id: 32fee0fd22de994b728a78745b7ff31756d15418